Minor balance shifts, added friendly torpedoes
Made some minor balance shifts to the ships, and added friendly torpedoes. Also removed the possibility for torpedoes to deal friendly fire damage - this was determined to be unfun, especially when it happened with the player's own torpedoes.

Friendly torpedoes still need tweaking to enable or disable them, though right now with free fire not working for them, the decision to target lock an enemy is a de facto declaration to fire torpedoes.